Refactor Books pages to Pages folder
