+++ LibraryApplication/LibraryApplication/Pages/Books/Add.aspx
@@ -1,4 +1,4 @@
-﻿<%@ Page Title="" Language="C#" MasterPageFile="~/Site.Master" AutoEventWireup="true" CodeBehind="AddBook.aspx.cs" Inherits="LibraryApplication.AddBook" %>
+﻿<%@ Page Title="" Language="C#" MasterPageFile="~/Site.Master" AutoEventWireup="true" CodeBehind="Add.aspx.cs" Inherits="LibraryApplication.Pages.Books.AddBook" %>
 
 <asp:Content ID="Content2" ContentPlaceHolderID="Main" runat="server">
     <fieldset>
