+++ LibraryApplication/LibraryApplication/Pages/Books/Edit.aspx
@@ -1,4 +1,4 @@
-﻿<%@ Page Title="" Language="C#" MasterPageFile="~/Site.Master" AutoEventWireup="true" CodeBehind="EditBook.aspx.cs" Inherits="LibraryApplication.EditBook" %>
+﻿<%@ Page Title="" Language="C#" MasterPageFile="~/Site.Master" AutoEventWireup="true" CodeBehind="Edit.aspx.cs" Inherits="LibraryApplication.Pages.Books.Edit" %>
 
 <asp:Content ID="Content2" ContentPlaceHolderID="Main" runat="server">
     <div>
