+++ LibraryApplication/LibraryApplication/Pages/Books/Edit.aspx.cs
@@ -1,18 +1,16 @@
 ﻿using Library.Data;
+using LibraryApplication.Data;
 using System;
-using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
-using System.Linq;
-using System.Web;
 using System.Web.UI;
 using System.Web.UI.WebControls;
 
-namespace LibraryApplication
+namespace LibraryApplication.Pages.Books
 {
-    public partial class EditBook : System.Web.UI.Page
+    public partial class Edit : System.Web.UI.Page
     {
-        private const string BooksUrl = "~/Books.aspx";
+        private string BooksUrl = SitePages.GetUrl(LibraryPage.Books);
         private const string GetBookQuery = @"
             SELECT 
                 Book.Id AS BookId,
